// src/app/shared/model/responsibility-center.ts
/**
 * Created by 15050978 on 12/21/2016.
 */
export class ResponsibilityCenter {
  s_responsibilityCenterCode: string;
  s_responsibilityCenterRealCode: string;
  s_responsibilityCenterAlias: string;
  s_responsibilityCenterName: string;
  s_responsibilityCenterAddress: string;
  s_responsibilityCenterPhone: string;
  s_responsibilityCenterFax: string;
  s_responsibilityCenterTaxCode: string;
  s_responsibilityCenterNpwp: string;
  s_responsibilityCenterEmail: string;
  d_responsibilityCenterStartDate: Date;
  d_responsibilityCenterSellingDate: Date;
  d_responsibilityCenterUpdatetableSellingDate: Date;
  d_responsibilityCenterPeriod: Date;
  s_responsibilityCenterServerLocalName: string;
  s_responsibilityCenterDBLocalName: string;
  i_responsibilityCenterCreateID: number;
  d_responsibilityCenterCreateDate: Date;
  i_responsibilityCenterDeleteID: number;
  d_responsibilityCenterDeleteDate: Date;
  i_responsibilityCenterUpdateID: number;
  s_responsibilityCenterActive: string;
}